requirements.txt
transformers
pinecone-client
torch
selenium
pandas
beautifulsoup4
nltk